Core run actions live on the runs surface
cancel and retry are run actions the frontend depends on, so they move from
the MCP gateway to druks/api/runs.py beside resume, keeping their operation
ids, docstrings, and wire shapes. Each carries the "agent" tag, which is all
the MCP derivation reads — tools/list still serves the same six tools with
unchanged annotations and schema constraints (fastmcp matches tags as a
subset). The gateway keeps what exists only for the tool surface: gates,
agent calls, usage.

The AgentApiError base and the run-action errors (RunNotFound, RunNotActive,
RunNotFailed, SubjectBusy) move to druks/api/exceptions.py so the runs
surface and the app's error handler no longer import from the protocol
adapter; gate and agent-call errors stay in the gateway, subclassing the api
base.
