+++ backend/druks/mcp/gateway/services.py
@@ -1,6 +1,7 @@
 from datetime import UTC, datetime, timedelta
 
 from druks.accounts.models import Account
+from druks.api.exceptions import RunNotFound
 from druks.core.utils.time import operator_local_day
 from druks.database import db_session
 from druks.durable.enums import RunState
@@ -28,7 +29,7 @@ _HISTORY_POINTS = 8
 def get_gate(run_id: str) -> schemas.GateResponse:
     run = Run.get(run_id)
     if not run:
-        raise exceptions.RunNotFound(run_id)
+        raise RunNotFound(run_id)
     if run.state != RunState.PARKED.value:
         raise exceptions.GateNotOpen(run_id)
     ask = run.input_request
@@ -48,7 +49,7 @@ async def answer_gate(
 ) -> schemas.GateAnswerResponse:
     run = Run.get(run_id)
     if not run:
-        raise exceptions.RunNotFound(run_id)
+        raise RunNotFound(run_id)
     db_session().expire(run)  # the receipt/park comparison must read fresh
     if run.answer_parked_at == parked_at:
         return schemas.GateAnswerResponse(
@@ -83,34 +84,6 @@ def get_agent_call(call_id: str) -> schemas.AgentCallDetailResponse:
         stderr=read_slice(layout.stderr, offset=-_STDERR_TAIL_BYTES, limit=_STDERR_TAIL_BYTES).text,
         artifact=_artifact_content(Artifact.get_for_call(call.id)),
     )
-
-
-async def cancel_run(run_id: str, *, reason: str) -> schemas.CancelRunResponse:
-    run = Run.get(run_id)
-    if not run:
-        raise exceptions.RunNotFound(run_id)
-    if run.state == RunState.CANCELLED.value:
-        return schemas.CancelRunResponse(run_id=run.id, result="already_cancelled")
-    if not run.is_active:
-        raise exceptions.RunNotActive(run_id)
-    await run.cancel(failure=reason)
-    return schemas.CancelRunResponse(run_id=run.id, result="cancelled")
-
-
-async def retry_run(run_id: str) -> schemas.RetryRunResponse:
-    run = Run.get(run_id)
-    if not run:
-        raise exceptions.RunNotFound(run_id)
-    if run.state != RunState.FAILED.value:
-        raise exceptions.RunNotFailed(run_id)
-
-    subject = run.subject
-    if subject:
-        latest = Run.get_latest_for_subject(subject["type"], subject["id"])
-        if latest and latest.is_active:
-            raise exceptions.SubjectBusy(latest.id)
-
-    return schemas.RetryRunResponse(run_id=await run.retry())
 
 
 def _artifact_content(artifact: Artifact | None) -> schemas.ArtifactContent | None:
